+++ src/main/java/com/example/projects/airBnb/service/RoomServiceImp.java
@@ -3,13 +3,16 @@ package com.example.projects.airBnb.service;
 import com.example.projects.airBnb.dto.RoomDto;
 import com.example.projects.airBnb.entity.Hotel;
 import com.example.projects.airBnb.entity.Room;
+import com.example.projects.airBnb.entity.User;
 import com.example.projects.airBnb.exception.ResourceNotFoundException;
+import com.example.projects.airBnb.exception.UnAuthorizedException;
 import com.example.projects.airBnb.repository.HotelRepository;
 import com.example.projects.airBnb.repository.RoomRepository;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -28,6 +31,12 @@ public class RoomServiceImp implements RoomService{
     public RoomDto createNewRoom(Long hotelId, RoomDto roomDto) {
         log.info("Create a new room in hotel with Id: {}", hotelId);
         Hotel hotel = hotelRepository.findById(hotelId).orElseThrow(() -> new ResourceNotFoundException("Hotel not found"));
+
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(user.equals(hotel.getOwner())){
+            throw new UnAuthorizedException("This user does not own this hotel with id: "+hotelId);
+        }
+
         Room room = modelMapper.map(roomDto, Room.class);
         room.setHotel(hotel);
         roomRepository.save(room);
@@ -41,6 +50,12 @@ public class RoomServiceImp implements RoomService{
     public List<RoomDto> getAllRoomsInHotel(Long hotelId) {
         log.info("getting all rooms");
         Hotel hotel = hotelRepository.findById(hotelId).orElseThrow(() -> new ResourceNotFoundException("Hotel not found"));
+
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(user.equals(hotel.getOwner())){
+            throw new UnAuthorizedException("This user does not own this hotel"+hotelId);
+        }
+
         return hotel.getRooms()
                 .stream()
                 .map((element)->modelMapper.map(element,RoomDto.class)).collect(Collectors.toList());
@@ -59,6 +74,12 @@ public class RoomServiceImp implements RoomService{
         log.info("deleting the room with Id: {}", roomId);
         boolean exists = roomRepository.existsById(roomId);
         Room room = roomRepository.findById(roomId).orElseThrow(() -> new ResourceNotFoundException("Room not found"));
+
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(user.equals(room.getHotel().getOwner())){
+            throw new UnAuthorizedException("This user does not own this room with id"+roomId);
+        }
+
         inventoryService.deleteFutureInventories(room);
         roomRepository.deleteById(roomId);
     }
